Replace tests to admin part
