+++ app/controllers/admin/ways_controller.rb
