+++ app/views/admin/tests/_scenario.html.slim
